requirements.txt
mne==1.6.1
numpy==1.26.4
scipy==1.13.0
tensorflow-macos==2.16.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
starlette==0.37.2
pandas==2.2.2
